# specs/011-project-client-access-control/project-client-access-control.openapi.yaml
openapi: 3.0.3
info:
  title: iTrack Project Client Access Control API
  version: 0.1.0
  description: Project-scoped client organization membership, invitations, verified domains, and trusted-domain policy.
servers:
  - url: /api
security:
  - sanctumCookie: []
paths:
...
  /client-organizations/{clientOrganization}/trusted-domain-policy:
    patch:
      summary: Update trusted-domain policy
      operationId: updateTrustedDomainPolicy
      tags: [Client Organizations]
      parameters:
        - $ref: "#/components/parameters/ClientOrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trusted_domain_policy]
              properties:
                trusted_domain_policy:
                  $ref: "#/components/schemas/TrustedDomainPolicy"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientOrganizationResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sanctumCookie:
      type: apiKey
      in: cookie
      name: laravel_session
  parameters:
    ClientOrganizationId:
      name: clientOrganization
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
...
  schemas:
    TrustedDomainPolicy:
      type: string
      enum: [domain_auto_approve, invitation_only, manual_approval]
...
    ClientOrganization:
      type: object
      required: [id, name, slug, trusted_domain_policy, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        trusted_domain_policy:
          $ref: "#/components/schemas/TrustedDomainPolicy"
        status:
          type: string
          enum: [active, suspended, archived]
...
    ClientOrganizationResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ClientOrganization"
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string